+++ components/membership/membership-client-impl/build.gradle
@@ -3,7 +3,7 @@ plugins {
     id 'corda.common-library'
 }
 
-description 'Membership HTTP RPC implementation'
+description 'Membership RPC Ops Client implementation'
 
 dependencies {
     compileOnly "org.osgi:osgi.annotation"
@@ -11,11 +11,15 @@ dependencies {
 
     implementation platform("net.corda:corda-api:$cordaApiVersion")
     implementation "net.corda.kotlin:kotlin-stdlib-jdk8-osgi"
+    implementation "net.corda:corda-avro-schema"
     implementation "net.corda:corda-base"
+    implementation "net.corda:corda-config-schema"
+    implementation "net.corda:corda-topic-schema"
 
-    implementation project(":libs:http-rpc:http-rpc")
+    implementation project(":components:configuration:configuration-read-service")
     implementation project(":libs:lifecycle:lifecycle")
     implementation project(":libs:membership:membership-http-rpc")
+    implementation project(":libs:messaging:messaging")
 
     testImplementation "org.mockito.kotlin:mockito-kotlin:$mockitoKotlinVersion"
     testImplementation "org.jetbrains.kotlin:kotlin-test:$kotlinVersion"
