+++ components/membership/membership-client-impl/src/test/kotlin/net/corda/membership/impl/client/lifecycle/MembershipRpcOpsClientLifecycleHandlerTest.kt
@@ -1,4 +1,4 @@
-package net.corda.membership.impl.httprpc.lifecycle
+package net.corda.membership.impl.client.lifecycle
 
 import net.corda.configuration.read.ConfigChangedEvent
 import net.corda.configuration.read.ConfigurationReadService
@@ -13,7 +13,7 @@ import net.corda.lifecycle.RegistrationHandle
 import net.corda.lifecycle.RegistrationStatusChangeEvent
 import net.corda.lifecycle.StartEvent
 import net.corda.lifecycle.StopEvent
-import net.corda.membership.impl.httprpc.MembershipRpcOpsClientImpl
+import net.corda.membership.impl.client.MembershipRpcOpsClientImpl
 import net.corda.messaging.api.publisher.RPCSender
 import net.corda.messaging.api.publisher.factory.PublisherFactory
 import net.corda.messaging.api.subscription.config.RPCConfig
