+++ settings.gradle
@@ -121,6 +121,7 @@ include 'components:install:test-resources:system-cpk'
 include 'components:kafka-topic-admin'
 include 'components:link-manager'
 include 'components:membership:group-policy-impl'
+include 'components:membership:membership-client-impl'
 include 'components:membership:membership-group-read-impl'
 include 'components:membership:membership-http-rpc-impl'
 include 'components:membership:registration-static-network-impl'
